pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-ocr-system"
version = "0.1.0"
description = "Enterprise-grade Chinese OCR Solution powered by PaddleOCR"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Smart OCR Team"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Text Processing :: General",
]
keywords = ["ocr", "paddleocr", "chinese", "text-recognition", "image-processing"]

dependencies = [
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
gpu = [
    "paddlepaddle-gpu>=2.5.0",
]

[project.scripts]
smart-ocr = "smart_ocr.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/smart-ocr-system"
Documentation = "https://github.com/your-org/smart-ocr-system#readme"
Repository = "https://github.com/your-org/smart-ocr-system.git"
Issues = "https://github.com/your-org/smart-ocr-system/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # defaults
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["smart_ocr"]
skip = [".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/smart_ocr"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
